requirements.txt
Flask
deepface
pillow
numpy
pytesseract
transformers
torch
librosa
soundfile
opencv-python
tf-keras
protobuf==3.20.*
sentencepiece
json
requests